test: split runtime orchestrator CLI contracts
